pyproject.toml
[project]
name = "code-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.62.0",
    "mcp[cli]>=1.26.0",
    "pillow>=12.1.0",
    "playwright>=1.58.0",
    "pywinauto>=0.6.9",
]
